Use mitm proxy library to support https
The previous library wasn't a man-in-the-middle proxy so it didn't
terminate https connections.  This meant that https requests wouldn't
get the proper headers added to them so the proxy was effectively only
usable with http.

This library actually terminates the connection, so it can change the
headers even when the client and server use https.

No attempt has been made so far to address the invalid certificate issue
that the client will experience when using this proxy.